chore: resolve all compiler and clippy warnings across workspace
Apply cargo clippy suggestions (189→14 remaining):
- Replace .max().min() with .clamp() in engine evaluation
- Use iterators instead of indexing loops (bvh, streaming_lod, pointcloud)
- Remove redundant field names, unnecessary to_string(), unnecessary parens
- Add is_empty() to ShaderVariantCache, rename HaltonSequence::next→next_halton
- Rename Group::add→add_shape, add #[allow] for dead_code/style warnings
- Suppress too_many_arguments/type_complexity in crate lib.rs where needed

+++ crates/rc3d-mesh/src/tessellate.rs
@@ -105,7 +105,7 @@ pub fn tessellate_sphere(radius: f32, slices: u32, stacks: u32) -> TriangleMesh 
     positions.push(Vec3::new(0.0, -radius, 0.0));
     texcoords.push([0.5, 1.0]);
 
-    let south_idx = (1 + ring_count * slices) as u32;
+    let south_idx = 1 + ring_count * slices;
 
     for j in 0..slices {
         let jp = (j + 1) % slices;
